Update TECHNICAL-STEERING-COMMITTEE.md
TSC members usually won't be in place when the Github site is launched (since they need to be approved by the PGB and the PGB may not convene until after launch). I think generic text like this will be better than a "Coming soon" message. We may want to instruct people interested in serving on the TSC to contact the PGB and not staff, though.